pagina principal

+++ db/migrate/20140727184832_sorcery_core.rb
@@ -5,7 +5,7 @@ class SorceryCore < ActiveRecord::Migration
       t.string :email,            :null => false
       t.string :crypted_password, :null => false
       t.string :salt,             :null => false
-      t.byte :level,            :null => false
+      t.decimal :level,            :null => false
       t.timestamps
     end
